==> Code/07_rasters_final.R <==
library(raster); library(ggplot2); library(sf)



##### current fire prob - max #####

# list current fire prob rasters without SE rasters
raster_filenames <- list.files('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/prediction_rasters_mean', pattern = 'monthly')
raster_filenames <- sapply(raster_filenames, function(x) gsub(pattern = '_SE', replacement = '', x))
raster_filenames <- raster_filenames[!duplicated(raster_filenames)]

# keep only filenames for islands/island groups we care about (remove duplicates)
island <- c('Hawaii', 'Oahu', 'Kauai', 'MauiCounty')
raster_filenames_split <- sapply(raster_filenames, function(x) strsplit(x, '_'))

# get unique groups
raster_groups <- unique(sapply(raster_filenames_split, function(x) x[[5]]))

# # for kauai, keep only KauaiOnlyModel
# kauai_files <- sapply(raster_filenames_split, function(x) x[[5]] == 'Kauai')
# kauai_files <- raster_filenames_split[kauai_files]
# kauai_files <- kauai_files[seq(2, 24, 2)]
# kauai_files <- names(kauai_files)
# 
# # get all non-kauai files
# nonkauai_files <- sapply(raster_filenames_split, function(x) x[[5]] != 'Kauai')
# nonkauai_files <- raster_filenames[nonkauai_files]
# 
# # merge kauai and non-kauai filenames
# raster_filenames <- c(nonkauai_files, kauai_files)
# rm(raster_filenames_split, kauai_files, nonkauai_files)

# get raster filename island group
raster_filenames_islandGroup <- strsplit(raster_filenames, '_')
raster_filenames_islandGroup <- sapply(raster_filenames_islandGroup, function(x) x[[5]])

# for each island group, load corresponding rasters
raster_list_dry <- list()
raster_list_wet <- list()
for(i in 1:length(raster_groups)){
  
  # load island group i's rasters
  rasters <- raster_filenames[raster_filenames_islandGroup == raster_groups[[i]]]
  rasters <- lapply(rasters, function(x) raster(paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/prediction_rasters_mean/', x)))
  
  # separate by wet and dry season
  rasters_wet <- rasters[c(11,12,1,2,3,4)]
  rasters_dry <- rasters[c(5,6,7,8,9,10)]
  rm(rasters); gc()
  
  # get wet and dry season maxes
  rasters_wet <- stack(rasters_wet); rasters_wet_mean <- calc(rasters_wet, max)
  rasters_dry <- stack(rasters_dry); rasters_dry_mean <- calc(rasters_dry, max)
  
  # # plot using ggplot
  # rasters_dry_mean_df <- as.data.frame(rasters_dry_mean, xy=TRUE)
  # rasters_dry_mean_df <- rasters_dry_mean_df[!is.na(rasters_dry_mean_df$layer),]
  # ggplot(data = rasters_dry_mean_df) + geom_tile(aes(x=x, y=y, color = layer, fill = layer))
  
  # save rasters to list
  raster_list_dry[[i]] <- rasters_dry_mean; names(raster_list_dry)[[i]] <- raster_groups[[i]]
  raster_list_wet[[i]] <- rasters_wet_mean; names(raster_list_wet)[[i]] <- raster_groups[[i]]
}
gc()



### for all rasters, remove risk pixels if landcover is NA ###

# load a landcover raster
raster_landcover <- raster("H:/My Drive/Projects/PICASC Land-to-sea/Data/Intermediate/Fire/00 Statewide landcover rasters resampled/HI_FracLC_bare_2016.tif")

# if landcover is NA, change risk to NA
for(i in 1:length(raster_list_dry)){
  
  # crop raster
  raster_landcover_cropped <- crop(raster_landcover, raster_list_dry[[i]])
  
  # change risk to NA if landcover is NA
  raster_list_dry[[i]][is.na(raster_landcover_cropped) & !is.na(raster_list_dry[[i]])] <- NA
  raster_list_wet[[i]][is.na(raster_landcover_cropped) & !is.na(raster_list_wet[[i]])] <- NA
  
}

rm(raster_landcover_cropped); gc()



### export rasters ###

# combine all wet and dry season rasters into statewide rasters
names(raster_list_dry) <- names(raster_list_wet) <- c('x', 'y')
raster_statewide_dry <- do.call(raster::merge, raster_list_dry)
raster_statewide_wet <- do.call(raster::merge, raster_list_wet)

# export island-level rasters
writeRaster(raster_statewide_dry, filename = 'H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_dry_statewide.tif', overwrite = TRUE)
writeRaster(raster_statewide_wet, filename = 'H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_wet_statewide.tif', overwrite = TRUE)

for(i in 1:length(raster_list_dry)){
  
  writeRaster(raster_list_dry[[i]], filename = paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_dry_', raster_groups[[i]], '.tif'), overwrite = TRUE)
  writeRaster(raster_list_wet[[i]], filename = paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_wet_', raster_groups[[i]], '.tif'), overwrite = TRUE)
  
}




##### deltas #####

# load delta files (significant deltas)
delta_files <- list.files('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/prediction_rasters_deltas/delta_rasters_with_significance', pattern = '.tif')

# separate files by model and island, find all unique model-island combinations
delta_files_separated <- strsplit(delta_files, '_')
delta_modelsAndIslands <- sapply(delta_files_separated, function(x) paste(x[1:10], collapse = '_'))
delta_modelsAndIslands <- delta_modelsAndIslands[!duplicated(delta_modelsAndIslands)]
rm(delta_files_separated)

# for each model-island, get create mean delta raster by season
raster_list_dry <- list()
raster_list_wet <- list()
for(i in 1:length(delta_modelsAndIslands)){
  
  # list then load monthly raster names for the model-island combo i
  rasters <- grep(delta_modelsAndIslands[[i]], delta_files, value=TRUE)
  rasters <- lapply(rasters, function(x) raster(paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/prediction_rasters_deltas/delta_rasters_with_significance/', x)))
  
  # separate by wet and dry season
  rasters_wet <- rasters[c(11,12,1,2,3,4)]
  rasters_dry <- rasters[c(5,6,7,8,9,10)]
  rm(rasters); gc()
  
  # get wet and dry season maxes
  rasters_wet <- stack(rasters_wet); rasters_wet_mean <- calc(rasters_wet, max)
  rasters_dry <- stack(rasters_dry); rasters_dry_mean <- calc(rasters_dry, max)
  
  # # plot using ggplot
  # rasters_dry_mean_df <- as.data.frame(rasters_dry_mean, xy=TRUE)
  # rasters_dry_mean_df <- rasters_dry_mean_df[!is.na(rasters_dry_mean_df$layer),]
  # ggplot(data = rasters_dry_mean_df) + geom_tile(aes(x=x, y=y, color = layer, fill = layer))
  
  # save rasters to list
  raster_list_dry[[i]] <- rasters_dry_mean; names(raster_list_dry)[[i]] <- delta_modelsAndIslands[[i]]
  raster_list_wet[[i]] <- rasters_wet_mean; names(raster_list_wet)[[i]] <- delta_modelsAndIslands[[i]]
  
  gc()
  
}
gc()

# # combine all wet and dry season rasters into statewide rasters
# names(raster_list_dry) <- names(raster_list_wet) <- c('x', 'y')
# raster_statewide_dry <- do.call(raster::merge, raster_list_dry)
# raster_statewide_wet <- do.call(raster::merge, raster_list_wet)
# 
# # export rasters
# writeRaster(raster_statewide_dry, filename = 'H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/mean_delta_dry_statewide.tif', overwrite = TRUE)
# writeRaster(raster_statewide_wet, filename = 'H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/mean_delta_wet_statewide.tif', overwrite = TRUE)

for(i in 1:length(raster_list_dry)){
  
  writeRaster(raster_list_dry[[i]], filename = paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/', delta_modelsAndIslands[[i]], '_dry.tif'), overwrite = TRUE)
  writeRaster(raster_list_wet[[i]], filename = paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/', delta_modelsAndIslands[[i]], '_wet.tif'), overwrite = TRUE)
  
}




##### create annual max #####

# get raster names
list_seasonalRasters <-
  list.files('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/',
             pattern = 'max_current', full.names = TRUE)

# for each island, save max annual
for(i in 1:length(island)){
  
  # get rasters for island i
  list_ras_names <- list_seasonalRasters[grep(island[[i]], list_seasonalRasters)]
  list_ras <- lapply(list_ras_names, raster)
  
  # get max annual risk
  ras_maxAnnual <- calc(stack(list_ras), max)
  
  # write raster
  writeRaster(ras_maxAnnual,
              filename = paste0('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_annual_',
                                island[[i]], '.tif'),
              overwrite = TRUE)
}

rm(list_seasonalRasters, i, list_ras, list_ras_names, ras_maxAnnual)
gc()

# combine into statewide raster and export
list_ras_names <-
  list.files('H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/',
             pattern = 'max_current_fire_prob_annual_',
             full.names = TRUE)
list_ras <- lapply(list_ras_names, raster)
raster_statewide_annual <- do.call(raster::merge, list_ras)
writeRaster(raster_statewide_annual,
            filename = 'H:/My Drive/Projects/PICASC Land-to-sea/Data/Processed/Fire/rasters_final/max_current_fire_prob_annual_statewide.tif',
            overwrite = TRUE)
